user delete check added

+++ build/controllers/userController.js
@@ -1,9 +1,11 @@
 import { pool } from '../db/DBConnect.js';
 const deleteUser = async (req, res) => {
-    // TODO data validation against DB types
     const { email } = req.body;
     if (!email)
         return res.status(400).json({ message: 'email required' });
+    const userEmail = req.userEmail;
+    if (email !== userEmail)
+        return res.sendStatus(403);
     // deleting user
     try {
         const result = await pool.execute('DELETE FROM users WHERE email = ?', [email]);
